requirements.txt
tiktoken>=0.12.0
pdfplumber>=0.11.9
sentence-transformers
scikit-learn
numpy
faiss-cpu
tqdm
llama-cpp-python